[ADD] replace active with recording and stop_recording
When only using state.active to determine the recording state it was
impossible to know whether the recording would stop at the next split.
Instead of active add recording and stop_recording. stop_recording act
as a signal that tells the recording loop to stop. recording is the
actual recording state.